Refactor the structure
